refactor(agents): one draft composer for the home page and the agent page
The agent page's Sessions tab now renders the same NewSessionComposer
as the home page, pinned to its agent (the agent slot shows the name
with no picker) and gaining the model badge, so a session started from
either place looks the same and can begin on a chosen model. The agent
page's own create-and-send flow is gone; the composer owns it, and a
user tool run from the palette creates the session the same way.
